Switch to new layout, replicating existing functionality
