--- cmd/api/healthcheck.go
@@ -1,21 +1,0 @@
-package main
-
-import (
-	"net/http"
-	"strconv"
-)
-
-func (app *application) healthcheckHandler(w http.ResponseWriter, r *http.Request) {
-	env := envelope{
-		"status": "available",
-		"system_info": map[string]string{
-			"debug":   strconv.FormatBool(app.config.Debug),
-			"version": "1.0.0",
-		},
-	}
-
-	err := app.writeJSON(w, http.StatusOK, env, nil)
-	if err != nil {
-		app.serverErrorResponse(w, r, err)
-	}
-}
