--- cmd/api/links_test.go
@@ -1,193 +1,0 @@
-package main
-
-import (
-	"bytes"
-	"fmt"
-	"github.com/danielmichaels/shortlink-go/internal/config"
-	"github.com/danielmichaels/shortlink-go/internal/data"
-	"github.com/danielmichaels/shortlink-go/internal/logger"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"testing"
-)
-
-func TestLinks(t *testing.T) {
-	cfg := config.AppConfig()
-	cfg.Db.DbName = "file::memory:?cache=shared"
-	log := logger.New(os.Stdout, logger.LevelInfo)
-	a := &application{
-		config: cfg,
-		logger: log,
-		models: data.NewMockModels(),
-	}
-
-	t.Run("GET request to /v1/links should return 405 response", func(t *testing.T) {
-		t.Helper()
-		r, err := http.NewRequest("GET", "/v1/links", nil)
-		if err != nil {
-			t.Fatal(err)
-		}
-		rr := httptest.NewRecorder()
-
-		h := a.routes()
-		h.ServeHTTP(rr, r)
-
-		if status := rr.Code; status != http.StatusMethodNotAllowed {
-			t.Errorf("handler returned the wrong status code. got %v want %v", rr.Code, http.StatusMethodNotAllowed)
-		}
-	})
-	t.Run("GET request to /v1/links/fake should return a 307 to Frontend 404 page", func(t *testing.T) {
-		t.Helper()
-		t.Log(cfg.Db)
-		hash := "fake"
-		r, err := http.NewRequest("GET", fmt.Sprintf("/v1/links/%s", hash), nil)
-		if err != nil {
-			t.Errorf("could not hit /v1/links/%s", hash)
-		}
-		rr := httptest.NewRecorder()
-
-		h := a.routes()
-		h.ServeHTTP(rr, r)
-
-		if status := rr.Code; status != http.StatusTemporaryRedirect {
-			t.Errorf("handler returned the wrong status code. got %v want %v", rr.Code, http.StatusTemporaryRedirect)
-		}
-
-		e, _ := rr.Result().Location()
-		expected := fmt.Sprintf("%s/404", a.config.Server.FrontendDomain)
-		result := fmt.Sprintf("%s://%s%s", e.Scheme, e.Host, e.Path)
-		if result != expected {
-			t.Errorf("result URL does not match expected URL. got %v want %v", result, expected)
-		}
-	})
-	t.Run("GET request to /v1/links/notfake should return a 307 to original URL", func(t *testing.T) {
-		t.Helper()
-		hash := "notfake"
-		r, err := http.NewRequest("GET", fmt.Sprintf("/v1/links/%s", hash), nil)
-		if err != nil {
-			t.Errorf("could not hit /v1/links/%s", hash)
-		}
-		rr := httptest.NewRecorder()
-
-		h := a.routes()
-		h.ServeHTTP(rr, r)
-
-		if status := rr.Code; status != http.StatusTemporaryRedirect {
-			t.Errorf("handler returned the wrong status code. got %v want %v", rr.Code, http.StatusTemporaryRedirect)
-		}
-
-		// Really this should return `http://test.com` but it's not for whatever reason.
-		expected := `/v1/links/test.com`
-		location, _ := rr.Result().Location()
-		if location.String() != expected {
-			t.Errorf("handler returned unexpected redirect. got %v want %v", location.String(), expected)
-		}
-	})
-	t.Run("POST a new link", func(t *testing.T) {
-		t.Helper()
-		link := `{"link":"test.com"}`
-		r, err := http.NewRequest("POST", "/v1/links", bytes.NewBufferString(link))
-		if err != nil {
-			t.Error("could not hit /v1/links")
-		}
-		rr := httptest.NewRecorder()
-
-		h := a.routes()
-		h.ServeHTTP(rr, r)
-
-		if status := rr.Code; status != http.StatusCreated {
-			t.Errorf("handler returned the wrong status code. got %v want %v", rr.Code, http.StatusOK)
-		}
-		// todo: FIX
-		// Can't get the `hash` and `short_link` values to align as its dynamically created
-		// on demand.
-		//expected := `{"link":{"id":0,"created_at":"0001-01-01T00:00:00Z","original_url":"test.com","hash":"sZ1TY61fD5i"},"short_url":"http://localhost:1988/sZ1TY61fD5i"}` + "\n"
-		//if rr.Body.String() != expected {
-		//	t.Errorf("handler returned unexpected body. got %v want %v", rr.Body.String(), expected)
-		//}
-	})
-	t.Run("GET request to /v1/links/{hash}/analytics should return 200 on success", func(t *testing.T) {
-		t.Helper()
-		hash := "notfake"
-		r, err := http.NewRequest("GET", fmt.Sprintf("/v1/links/%s/analytics", hash), nil)
-		if err != nil {
-			t.Errorf("could not GET /v1/links/%s/analytics", hash)
-		}
-		rr := httptest.NewRecorder()
-
-		h := a.routes()
-		h.ServeHTTP(rr, r)
-
-		if status := rr.Code; status != http.StatusOK {
-			t.Errorf("handler returned the wrong status code. got %v want %v", rr.Code, http.StatusOK)
-		}
-		expected := `{"analytics":[{"id":1,"ip_address":"1.1.1.1","user_agent":"test-agent","date_accessed":"2020-01-01T00:00:00Z"},{"id":1,"ip_address":"1.1.1.1","user_agent":"test-agent","date_accessed":"2020-01-01T00:00:00Z"}],"metadata":{"current_page":1,"page_size":20,"first_page":1,"last_page":1,"total_records":2}}` + "\n"
-		if rr.Body.String() != expected {
-			t.Errorf("handler returned unexpected body. got %v want %v", rr.Body.String(), expected)
-		}
-	})
-	t.Run("GET request to /v1/links/{hash}/analytics should return 200 but empty response structs if not found", func(t *testing.T) {
-		t.Helper()
-		hash := "fake"
-		r, err := http.NewRequest("GET", fmt.Sprintf("/v1/links/%s/analytics", hash), nil)
-		if err != nil {
-			t.Errorf("could not GET /v1/links/%s/analytics", hash)
-		}
-		rr := httptest.NewRecorder()
-
-		h := a.routes()
-		h.ServeHTTP(rr, r)
-
-		if status := rr.Code; status != http.StatusOK {
-			t.Errorf("handler returned the wrong status code. got %v want %v", rr.Code, http.StatusOK)
-		}
-		expected := `{"analytics":[],"metadata":{}}` + "\n"
-		t.Log(rr.Body.String())
-		if rr.Body.String() != expected {
-			t.Errorf("handler returned unexpected body. got %v want %v", rr.Body.String(), expected)
-		}
-	})
-	t.Run("GET request to /v1/links/{hash}/analytics?page=1&page_size=1 should return 200 but empty response structs if not found", func(t *testing.T) {
-		t.Helper()
-		hash := "pagination"
-		r, err := http.NewRequest("GET", fmt.Sprintf("/v1/links/%s/analytics?page=1&page_size=1", hash), nil)
-		if err != nil {
-			t.Errorf("could not GET /v1/links/%s/analytics", hash)
-		}
-		rr := httptest.NewRecorder()
-
-		h := a.routes()
-		h.ServeHTTP(rr, r)
-
-		if status := rr.Code; status != http.StatusOK {
-			t.Errorf("handler returned the wrong status code. got %v want %v", rr.Code, http.StatusOK)
-		}
-		expected := `{"analytics":[{"id":1,"ip_address":"1.1.1.1","user_agent":"test-agent","date_accessed":"2020-01-01T00:00:00Z"}],"metadata":{"current_page":1,"page_size":1,"first_page":1,"last_page":1,"total_records":1}}` + "\n"
-		t.Log(rr.Body.String())
-		if rr.Body.String() != expected {
-			t.Errorf("handler returned unexpected body. got %v want %v", rr.Body.String(), expected)
-		}
-	})
-	t.Run("GET request to /v1/links/{hash}/analytics?page=0&page_size=1 should return 422 and error due to validation failure", func(t *testing.T) {
-		t.Helper()
-		hash := "pagination"
-		r, err := http.NewRequest("GET", fmt.Sprintf("/v1/links/%s/analytics?page=0&page_size=1", hash), nil)
-		if err != nil {
-			t.Errorf("could not GET /v1/links/%s/analytics", hash)
-		}
-		rr := httptest.NewRecorder()
-
-		h := a.routes()
-		h.ServeHTTP(rr, r)
-
-		if status := rr.Code; status != http.StatusUnprocessableEntity {
-			t.Errorf("handler returned the wrong status code. got %v want %v", rr.Code, http.StatusOK)
-		}
-		expected := `{"error":{"page":"must be greater than zero"}}` + "\n"
-		t.Log(rr.Body.String())
-		if rr.Body.String() != expected {
-			t.Errorf("handler returned unexpected body. got %v want %v", rr.Body.String(), expected)
-		}
-	})
-}
