--- cmd/api/middleware.go
@@ -1,110 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"github.com/tomasen/realip"
-	"golang.org/x/time/rate"
-	"net/http"
-	"sync"
-	"time"
-)
-
-func (app *application) recoverPanic(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		defer func() {
-			if err := recover(); err != nil {
-				w.Header().Set("Connection", "close")
-				app.serverErrorResponse(w, r, fmt.Errorf("%s", err))
-			}
-		}()
-
-		next.ServeHTTP(w, r)
-	})
-}
-
-func (app *application) enableCORS(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Add("Vary", "Origin")
-
-		w.Header().Add("Vary", "Access-Control-Request-Method")
-
-		origin := r.Header.Get("Origin")
-
-		if origin != "" {
-			for i := range app.config.Server.AllowedOrigins {
-				if origin == app.config.Server.AllowedOrigins[i] {
-					w.Header().Set("Access-Control-Allow-Origin", origin)
-
-					if r.Method == http.MethodOptions && r.Header.Get("Access-Control-Request-Method") != "" {
-
-						w.Header().Set("Access-Control-Allow-Methods", "OPTIONS, PUT, PATCH, DELETE")
-						w.Header().Set("Access-Control-Allow-Headers", "Authorization, Content-Type")
-
-						w.WriteHeader(http.StatusOK)
-						return
-					}
-
-					break
-				}
-			}
-		}
-
-		next.ServeHTTP(w, r)
-	})
-}
-
-// rateLimit limits the number of requests by IP address.
-func (app *application) rateLimit(next http.Handler) http.Handler {
-	type client struct {
-		limiter  *rate.Limiter
-		lastSeen time.Time
-	}
-
-	var (
-		mu      sync.Mutex
-		clients = make(map[string]*client)
-	)
-
-	// Run a cleanup every minute
-	go func() {
-		for {
-			time.Sleep(time.Minute)
-			mu.Lock()
-			for ip, client := range clients {
-				if time.Since(client.lastSeen) > 3*time.Minute {
-					delete(clients, ip)
-				}
-			}
-			mu.Unlock()
-		}
-	}()
-
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if app.config.Limiter.Enabled {
-			ip := realip.FromRequest(r)
-			mu.Lock()
-
-			// Check to see if the IP address already exists in the map. If it doesn't
-			// then init a new rate limiter and add the IP address and limiter to
-			// the map.
-			if _, found := clients[ip]; !found {
-				clients[ip] = &client{
-					limiter: rate.NewLimiter(rate.Limit(app.config.Limiter.Rps), app.config.Limiter.Burst),
-				}
-			}
-
-			// Update client's last seen time.
-			clients[ip].lastSeen = time.Now()
-			// Test the ip with Allow() and if it needs limiting return 429's
-			if !clients[ip].limiter.Allow() {
-				mu.Unlock()
-				app.rateLimitExceededResponse(w, r)
-				return
-			}
-			mu.Unlock()
-		}
-		// Allow the request through.
-		next.ServeHTTP(w, r)
-	})
-
-}
