--- cmd/web/helpers.go
@@ -1,104 +1,0 @@
-package main
-
-import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"net/http"
-	"strings"
-)
-
-type envelope map[string]interface{}
-
-func (app *application) writeJSON(w http.ResponseWriter, status int, data envelope, headers http.Header) error {
-	js, err := json.Marshal(data)
-	if err != nil {
-		return err
-	}
-	js = append(js, '\n')
-
-	for key, value := range headers {
-		w.Header()[key] = value
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	_, _ = w.Write(js)
-	return nil
-}
-
-func (app *application) readJSON(w http.ResponseWriter, r *http.Request, dst interface{}) error {
-	maxBytes := 1_048_576
-	r.Body = http.MaxBytesReader(w, r.Body, int64(maxBytes))
-
-	dec := json.NewDecoder(r.Body)
-	dec.DisallowUnknownFields()
-
-	err := dec.Decode(dst)
-	if err != nil {
-		var syntaxError *json.SyntaxError
-		var unmarshallTypeError *json.UnmarshalTypeError
-		var invalidUnmarshallError *json.InvalidUnmarshalError
-
-		switch {
-		case errors.As(err, &syntaxError):
-			return fmt.Errorf("body contains badly-formed JSON (at character %d)", syntaxError.Offset)
-		case errors.Is(err, io.ErrUnexpectedEOF):
-			return errors.New("body contains badly-formed JSON")
-		case errors.As(err, &unmarshallTypeError):
-			if unmarshallTypeError.Field != "" {
-				return fmt.Errorf("body contains incorrect JSON type for field %q", unmarshallTypeError.Field)
-			}
-			return fmt.Errorf("body contains incorrect JSON type (at character %d)", unmarshallTypeError.Offset)
-		case errors.Is(err, io.EOF):
-			return errors.New("body must not be empty")
-		case strings.HasPrefix(err.Error(), "json: unknown field "):
-			fieldName := strings.TrimPrefix(err.Error(), "json: unknown field ")
-			return fmt.Errorf("body contains unknown key %s", fieldName)
-		case err.Error() == "http: request body too large":
-			return fmt.Errorf("body must not be larger than %d bytes", maxBytes)
-		case errors.As(err, &invalidUnmarshallError):
-			panic(err)
-		default:
-			return err
-		}
-	}
-
-	err = dec.Decode(&struct{}{})
-	if err != io.EOF {
-		return errors.New("body must contain a single JSON value")
-	}
-	return nil
-}
-
-// addDefaultData is a helper which will pre-fill the templateData struct with
-// default information that is used across several templates.
-func (app *application) addDefaultData(td *templateData, r *http.Request) *templateData {
-	if td == nil {
-		td = &templateData{}
-	}
-	td.AppUrl = app.config.Server.FrontendDomain
-	td.Names.AppName = app.config.Names.AppName
-	td.Names.Twitter = app.config.Names.TwitterAccount
-	td.Names.Github = app.config.Names.GithubAccount
-	td.Names.Plausible = app.config.Names.PlausibleAccount
-	return td
-}
-
-// render is a template rendering helper. It uses a template cache to speed up delivery of templates
-func (app *application) render(w http.ResponseWriter, r *http.Request, name string, td *templateData) {
-	ts, ok := app.template[name]
-	if !ok {
-		http.Error(w, "Template does not exist", 500)
-		return
-	}
-	buf := new(bytes.Buffer)
-	err := ts.Execute(buf, app.addDefaultData(td, r))
-	if err != nil {
-		http.Error(w, err.Error(), 500)
-		return
-	}
-	buf.WriteTo(w)
-}
