+++ internal/cmd/serve.go
@@ -3,9 +3,11 @@ package cmd
 import (
 	"context"
 	"github.com/danielmichaels/shortlink-go/internal/config"
+	"github.com/danielmichaels/shortlink-go/internal/data"
 	zlog "github.com/danielmichaels/shortlink-go/internal/logger"
 	"github.com/danielmichaels/shortlink-go/internal/server"
 	"github.com/danielmichaels/shortlink-go/internal/store"
+	"github.com/danielmichaels/shortlink-go/internal/templates"
 	"github.com/spf13/cobra"
 )
 
@@ -17,20 +19,20 @@ func ServeCmd(ctx context.Context) *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) error {
 			cfg := config.AppConfig()
 			logger := zlog.NewLogger("tars", cfg)
-			_, err := store.OpenDB(cfg)
-			//if err != nil {
-			//	logger.Fatal().Err(err).Msg("failed to open database. exiting")
-			//}
+			db, err := store.OpenDB(cfg)
+			if err != nil {
+				logger.Fatal().Err(err).Msg("failed to open database. exiting")
+			}
 			logger.Info().Msg("database connection established")
-			//templateCache, err := templates.NewTemplateCache()
-			//if err != nil {
-			//	logger.Fatal().Err(err).Msg("failed to create a template cache")
-			//}
+			templateCache, err := templates.NewTemplateCache()
+			if err != nil {
+				logger.Fatal().Err(err).Msg("failed to create a template cache")
+			}
 			app := &server.Application{
-				Config: cfg,
-				Logger: logger,
-				//models:   data.NewModels(db),
-				//template: templateCache,
+				Config:   cfg,
+				Logger:   logger,
+				Models:   data.NewModels(db),
+				Template: templateCache,
 			}
 
 			err = app.Serve(ctx)
