+++ internal/server/helpers.go
@@ -1,20 +1,18 @@
-package main
+package server
 
 import (
+	"bytes"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/danielmichaels/shortlink-go/internal/validator"
 	"io"
 	"net/http"
-	"net/url"
-	"strconv"
 	"strings"
 )
 
-type envelope map[string]interface{}
+type M map[string]interface{}
 
-func (app *application) writeJSON(w http.ResponseWriter, status int, data envelope, headers http.Header) error {
+func (app *Application) writeJSON(w http.ResponseWriter, status int, data M, headers http.Header) error {
 	js, err := json.Marshal(data)
 	if err != nil {
 		return err
@@ -31,7 +29,7 @@ func (app *application) writeJSON(w http.ResponseWriter, status int, data envelo
 	return nil
 }
 
-func (app *application) readJSON(w http.ResponseWriter, r *http.Request, dst interface{}) error {
+func (app *Application) readJSON(w http.ResponseWriter, r *http.Request, dst interface{}) error {
 	maxBytes := 1_048_576
 	r.Body = http.MaxBytesReader(w, r.Body, int64(maxBytes))
 
@@ -75,37 +73,32 @@ func (app *application) readJSON(w http.ResponseWriter, r *http.Request, dst int
 	return nil
 }
 
-// readInt() reads string value from the query string and converts it to an integer
-// before returning. If no matching key is found it returns the provided default
-// value. If the value couldn't be converted to an int, then we record an error
-// message in the provided Validator instance.
-func (app *application) readInt(qs url.Values, key string, defaultValue int, v *validator.Validator) int {
-	s := qs.Get(key)
-
-	if s == "" {
-		return defaultValue
+// addDefaultData is a helper which will pre-fill the templateData struct with
+// default information that is used across several templates.
+func (app *Application) addDefaultData(td *templateData, r *http.Request) *templateData {
+	if td == nil {
+		td = &templateData{}
 	}
-
-	// Try to convert the value to an int. If this fails, add an error message to
-	// the validator instance and return the default value.
-	i, err := strconv.Atoi(s)
-	if err != nil {
-		v.AddError(key, "must be an integer value")
-		return defaultValue
-	}
-	return i
+	td.AppUrl = app.Config.Server.FrontendDomain
+	td.Names.AppName = app.Config.Names.AppName
+	td.Names.Twitter = app.Config.Names.TwitterAccount
+	td.Names.Github = app.Config.Names.GithubAccount
+	td.Names.Plausible = app.Config.Names.PlausibleAccount
+	return td
 }
 
-// readString() returns a string value from the query string, or the provided
-// default value if no matching key is found.
-func (app *application) readString(qs url.Values, key string, defaultValue string) string {
-	// Extract the value for a given key from the query string. If no key exists
-	// this will return the empty string "".
-	s := qs.Get(key)
-
-	// If no key exists (or value is empty) then return the default value.
-	if s == "" {
-		return defaultValue
+// render is a template rendering helper. It uses a template cache to speed up delivery of templates
+func (app *Application) render(w http.ResponseWriter, r *http.Request, name string, td *templateData) {
+	ts, ok := app.Template[name]
+	if !ok {
+		http.Error(w, "Template does not exist", 500)
+		return
 	}
-	return s
+	buf := new(bytes.Buffer)
+	err := ts.Execute(buf, app.addDefaultData(td, r))
+	if err != nil {
+		http.Error(w, err.Error(), 500)
+		return
+	}
+	_, _ = buf.WriteTo(w)
 }
