+++ internal/server/server.go
@@ -5,7 +5,9 @@ import (
 	"errors"
 	"fmt"
 	"github.com/danielmichaels/shortlink-go/internal/config"
+	"github.com/danielmichaels/shortlink-go/internal/data"
 	"github.com/rs/zerolog"
+	"html/template"
 	"net/http"
 	"os"
 	"os/signal"
@@ -18,6 +20,24 @@ type Application struct {
 	Config *config.Conf
 	Logger zerolog.Logger
 	//DB     *store.Queries
+	Models   data.Models
+	Template map[string]*template.Template
+}
+
+type templateData struct {
+	Title     string
+	AppUrl    string
+	Names     userTemplateData
+	Link      *data.Link
+	Analytics []*data.Analytics
+	Metadata  data.Metadata
+}
+
+type userTemplateData struct {
+	AppName   string
+	Github    string
+	Twitter   string
+	Plausible string
 }
 
 func (app *Application) Serve(ctx context.Context) error {
