+++ package.json
@@ -1,8 +1,8 @@
 {
   "scripts": {
-    "tailwind": "tailwindcss -i assets/static/css/main.css -o ui/static/css/theme.css -m",
-    "alpine": "npx esbuild index.js --outfile=ui/static/js/bundle.js --bundle --minify",
-    "alpine-dev": "npx esbuild index.js --outfile=ui/static/js/bundle.js --bundle"
+    "tailwind": "tailwindcss -i assets/static/css/main.css -o assets/static/css/theme.css -m",
+    "alpine": "npx esbuild index.js --outfile=assets/static/js/bundle.js --bundle --minify",
+    "alpine-dev": "npx esbuild index.js --outfile=assets/static/js/bundle.js --bundle"
   },
   "devDependencies": {
     "@tailwindcss/aspect-ratio": "^0.4.2",
